requirements.txt
Keras==2.4.3
matplotlib==3.4.2
streamlit==0.82.0
opencv_python==4.1.1.26
tensorflow==2.4.1
numpy==1.19.5
Pillow==8.2.0
